% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2x2.R
\name{draw_2x2}
\alias{draw_2x2}
\title{Visualise a 2x2 matrix}
\usage{
draw_2x2(matrix)
}
\arguments{
\item{matrix}{The 2x2 matrix to visualise}
}
\value{
a ggplot2 object
}
\description{
Visualise a 2x2 matrix
}
