% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2x2.R
\name{squares2x2}
\alias{squares2x2}
\title{Visualise a 2x2 matrix}
\usage{
squares2x2(
  matrix,
  labels = NULL,
  colors = c("#74ADD1", "#D73027", "#D73027", "#74ADD1"),
  labels_size = 4
)
}
\arguments{
\item{matrix}{The 2x2 matrix to visualise}

\item{labels}{The labels to display}

\item{colors}{Colors to use}

\item{labels_size}{Size of the labels}
}
\value{
a ggplot2 object
}
\description{
Visualise a 2x2 matrix
}
