% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis2x2-package.R
\docType{package}
\name{vis2x2-package}
\alias{vis2x2}
\alias{vis2x2-package}
\title{vis2x2: Visualisation of Two by Two Tables for Diagnostic Test Evaluation}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Paolo Eusebi \email{paoloeusebi@gmail.com}
}

}
\keyword{internal}
