% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2x2.R
\name{waffell2x2}
\alias{waffell2x2}
\title{Visualize a 2x2 matrix with waffell}
\usage{
waffell2x2(
  matrix,
  n = 1000,
  colors = c("#74ADD1", "#D73027", "#D73027", "#74ADD1")
)
}
\arguments{
\item{matrix}{The 2x2 matrix to visualize}

\item{colors}{Colors to use}

\item{labels}{The labels to display}

\item{labels_size}{Size of the labels}
}
\value{
a ggplot2 object
}
\description{
Visualize a 2x2 matrix with waffell
}
