% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2x2.R
\name{squares2x2}
\alias{squares2x2}
\title{Visualize a 2x2 matrix}
\usage{
squares2x2(x, labels = NULL, colors = "white", labels_size = 4)
}
\arguments{
\item{x}{A named vector to visualize}

\item{labels}{The labels to display clockwise from topleft}

\item{colors}{Colors to use from topleft}

\item{labels_size}{Size of the labels}
}
\value{
a ggplot2 object
}
\description{
Visualize a 2x2 matrix
}
